Now we can inline setting the metadata info in CreateColumnChunk.
